// src/enums/CultivationType.ts
export enum CultivationType {
    BLE = "blé",
    ORGE = "orge",
    AVOINE = "avoine",
    CANOLA = "canola",
    SOJA = "soja",
    RAISIN = "raisin",
    OLIVE = "olive",
    POMME_DE_TERRE = "pommeDeTerre",
    BETTERAVE = "betterave",
    COTTON = "cotton",

    MAIS = "mais",
    TOURNESOL = "tournesol",
    CANNE_A_SUCRE = "canneASucre",
    PEUPLIER = "peuplier",

    LEGUMES = "legumes",
    EPINARD = "epinard",
    POIS = "pois",
    HARICOTS_VERTS = "haricotsVerts",
    HERBES = "herbes",
    CACAO = "cacao",
    SERRE = "serre",
}